Added priority feature.
Added priority feature to the app, where each todo item can be
prioritized. The prioritized state is stored in the model, and a toggle
function for the priority was added as well. A prioritize button was added
in the html template. In the todo-view file the priority feature was
hooked up with functions to toggle the priority, their event bindings and
changes to the render function to facilitate the feature.